Implement create id/seq file function.
It's possible that the sequences of several fasta ids are the same. To
avoid unnecessary sequence comaprsion, the function parses the fasta
file and builds seq-to-ids mapping list.

Once have these files, the compare function can read seq file to compare
to reduce the comparasion time.

In the final process, the program needs id file to find the mapping.